app/requirements.txt
python-telegram-bot
python-dotenv
google-genai
playwright
pywinauto
pyautogui
